requirements.txt
weaviate_client==2.0.2
